requirements.txt
Click==7.0
comtypes==1.1.7
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
py-getch==1.0.1
pymongo==3.8.0
pynput==1.4.2
pypiwin32==223
pywin32==224
pywinauto==0.6.6
six==1.12.0
visitor==0.1.3
Werkzeug==0.15.2
WMI==1.4.9
